% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fagomr}
\alias{fagomr}
\title{Fagområdeinndeling}
\format{An object of class \code{list} of length 18.}
\usage{
fagomr
}
\description{
Oversikt over hvilke kvalitetsregistere som hører til under ulike fagområder
}
\keyword{datasets}
