% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KvalIndData}
\alias{KvalIndData}
\title{Kvalitetsindikatordata}
\format{Dataramme med ei rad per hendelse (operasjon/innleggelse/...)

\describe{
\item{Aar}{Årtall for hendelsen. format: YYYY}
\item{ShNavn}{Sykehusnavn. Dette er sykehusnavnet som benyttes i registeret. Det kan være registrert
hendelser fra flere enheter på ett og samme sykehus. Dette navnet kan da benyttes til å skille mellom
de ulike enhetene. Tekst}
\item{ReshId}{Den registrerende enhetens resh-id.}
\item{SykehusId}{Sykehusets organisasjonsnummer i Brønnøysundregisteret. Numerisk, 9 siffer}
\item{Variabel}{Resultatvariabel for kvalitetsindikatoren. I de fleste tilfeller vil dette være en
indikatorvariabel, dvs. ha verdiene 0 og 1. For noen kvalitetsindikatorer kan det være nødvendig å
benytte en eksakt verdi. F.eks. liggetid hvis man ønsker å benytte median liggetid som indikator.}
\item{kvalIndID}{ID for kvalitetsindikatoren. Benyttes til å koble indikatorbeskrivelser
til de observerte verdiene (eller omvendt...). Dvs. kobler sammen 'KvalIndData' og 'IndBeskr'.}
}}
\source{
Organisasjonsnumre og struktur kan finnes ved hjelp av Brønnøyaundregisteret: \url{https://w2.brreg.no/enhet/sok/}
}
\usage{
KvalIndData
}
\description{
"Hendelsesentydige" data som angir verdier for hver kvalitetsindikator.
}
\keyword{datasets}
