% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SykehusNavnStruktur}
\alias{SykehusNavnStruktur}
\title{Oversikt over sykehusstruktur og navn}
\format{
Dataramme med ei rad per sykehus og RHF, HF og sykehusnavn som variabler
\describe{
\item{OrgNrRHF}{RHF'ets organisasjonsnummer i Brønnøysundregisteret}
\item{RHF}{Regionalt Helseforetak}
\item{OrgNrHF}{HF'ets organisasjonsnummer i Brønnøysundregisteret}
\item{HF}{Helseforetak, navn fra Brønnøysundregisteret}
\item{Hfkortnavn}{Kortnavn for HF'et}
\item{OrgNrShus}{Sykehusets organisasjonsnummer i Brønnøysundregisteret}
\item{OrgNavnEnhetsreg}{Sykehusets navn i Brønnøysundregisteret}
\item{SykehusnavnLang}{Sykehusets navn "hos SKDE", dvs. navnet som har blitt benyttet i helseatlas.}
\item{SykehusNavn}{Sykehusets kortnavn}
\item{TaMed}{Indikator (0/1) som angir om sykehuset er vanlig å ha med i div. oversikter}
}
}
\source{
Organisasjonsnumre og struktur kan finnes ved hjelp av Brønnøyaundregisteret:
\url{https://w2.brreg.no/enhet/sok/}
}
\usage{
SykehusNavnStruktur
}
\description{
Datasett som inneholder oversikt over organisasjonsnummer og navn for RHF, HF og sykehus.
Organisasjonsnummer er hentet fra Brønnøysundregisteret. Navn fra Brønnøysundregisteret er også med, samt
"selvvalgte" kortnavn for både HF og sykehus.
}
\keyword{datasets}
